ironside: add initial IronSide SE support package
Rev: e6cc6ea5b7b85f31e892d20410f97dab1b24a331

Adds a new set of source files/scripts used to interact with
IronSide SE. The plan is that this will replace much of the
implementation that currently lives in zephyr/soc/nordic/ironside once
it is fully integrated.

The initial package includes:

* Serialization code for IronSide calls
* Boot report definition
* UICR definition
* Macros for encoding PERIPHCONF entries
* Addresses for certain IronSide SE defined structures in RAM
  (IPC buffers, boot reports, event reports)
* Macros definining the released IronSide SE versions, including
  a concept of "target" IronSide SE version which will enable
  compatibility with older IronSide SE versions.
* Python code for generating UICR and PERIPHCONF blobs
* Python code for validating a PERIPHCONF blob

For now, the code is developed and tested in a different repository and
will be copied into this repository on a regular basis. A git sha is
used to track the origin commit/tag in that repository, similarly to
other projects integrated in hal_nordic. The origin SHA is also
contained in ironside/git_sha.txt.

Signed-off-by: Jonathan Nilsen <jonathan.nilsen@nordicsemi.no>